+++ enterprise/integrations/gitlab/gitlab_manager.py
@@ -27,13 +27,13 @@ from server.auth.token_manager import TokenManager
 
 from openhands.core.logger import openhands_logger as logger
 from openhands.integrations.gitlab.gitlab_service import GitLabServiceImpl
+from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.integrations.provider import ProviderToken, ProviderType
 from openhands.server.types import (
     LLMAuthenticationError,
     MissingSettingsError,
     SessionExpiredError,
 )
-from openhands.storage.data_models.secrets import Secrets
 
 
 class GitlabManager(Manager[GitlabViewType]):
