+++ enterprise/storage/org_member_store.py
@@ -14,7 +14,7 @@ from storage.org_member import OrgMember
 from storage.user import User
 from storage.user_settings import UserSettings
 
-from openhands.storage.data_models.settings import Settings
+from openhands.app_server.settings.settings_models import Settings
 from openhands.utils.jsonpatch_compat import deep_merge
