+++ openhands/app_server/secrets/file_secrets_store.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 import json
 from dataclasses import dataclass
 
+from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.app_server.secrets.secrets_store import SecretsStore
 from openhands.core.config.openhands_config import OpenHandsConfig
 from openhands.storage import get_file_store
-from openhands.storage.data_models.secrets import Secrets
 from openhands.storage.files import FileStore
 from openhands.utils.async_utils import call_sync_from_async
