+++ openhands/app_server/secrets/secrets_router.py
@@ -12,8 +12,10 @@ from openhands.app_server.secrets.secrets_models import (
     CustomSecretCreate,
     CustomSecretPage,
     CustomSecretWithoutValue,
+    Secrets,
 )
 from openhands.app_server.secrets.secrets_store import SecretsStore
+from openhands.app_server.settings.settings_models import POSTProviderModel
 from openhands.app_server.utils.dependencies import get_dependencies
 from openhands.app_server.utils.models import EditResponse
 from openhands.integrations.provider import (
@@ -22,15 +24,11 @@ from openhands.integrations.provider import (
     ProviderType,
 )
 from openhands.integrations.utils import validate_provider_token
-from openhands.server.settings import (
-    POSTProviderModel,
-)
 from openhands.server.user_auth import (
     get_provider_tokens,
     get_secrets,
     get_secrets_store,
 )
-from openhands.storage.data_models.secrets import Secrets
 
 # Create router with /api/v1/secrets prefix
 router = APIRouter(
