+++ openhands/app_server/settings/file_settings_store.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 import json
 from dataclasses import dataclass
 
+from openhands.app_server.settings.settings_models import Settings
 from openhands.app_server.settings.settings_store import SettingsStore
 from openhands.core.config.openhands_config import OpenHandsConfig
 from openhands.storage import get_file_store
-from openhands.storage.data_models.settings import Settings
 from openhands.storage.files import FileStore
 from openhands.utils.async_utils import call_sync_from_async
