+++ openhands/server/settings.py
@@ -6,40 +6,21 @@
 # Unless you are working on deprecation, please avoid extending this legacy file and consult the V1 codepaths above.
 # Tag: Legacy-V0
 # This module belongs to the old V0 web server. The V1 application server lives under openhands/app_server/.
+#
+# Re-exports from new location for backward compatibility.
+# All models have been moved to openhands.app_server.settings.settings_models
 from __future__ import annotations
 
-from pydantic import (
-    BaseModel,
-    ConfigDict,
+from openhands.app_server.settings.settings_models import (
+    CustomSecretWithoutValueModel,
+    GETSettingsModel,
+    POSTProviderModel,
+    Settings,
 )
 
-from openhands.core.config.mcp_config import MCPConfig
-from openhands.integrations.provider import ProviderToken
-from openhands.integrations.service_types import ProviderType
-from openhands.storage.data_models.settings import Settings
-
-
-class POSTProviderModel(BaseModel):
-    """Settings for POST requests"""
-
-    mcp_config: MCPConfig | None = None
-    provider_tokens: dict[ProviderType, ProviderToken] = {}
-
-
-class GETSettingsModel(Settings):
-    """Settings with additional token data for the frontend"""
-
-    provider_tokens_set: dict[ProviderType, str | None] | None = (
-        None  # provider + base_domain key-value pair
-    )
-    llm_api_key_set: bool
-    search_api_key_set: bool = False
-
-    model_config = ConfigDict(use_enum_values=True)
-
-
-class CustomSecretWithoutValueModel(BaseModel):
-    """Custom secret model without value"""
-
-    name: str
-    description: str | None = None
+__all__ = [
+    'CustomSecretWithoutValueModel',
+    'GETSettingsModel',
+    'POSTProviderModel',
+    'Settings',
+]
