+++ openhands/server/user_auth/user_auth.py
@@ -14,13 +14,13 @@ from enum import Enum
 from fastapi import Request
 from pydantic import SecretStr
 
+from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.app_server.secrets.secrets_store import SecretsStore
+from openhands.app_server.settings.settings_models import Settings
 from openhands.app_server.settings.settings_store import SettingsStore
 from openhands.integrations.provider import PROVIDER_TOKEN_TYPE, ProviderHandler
 from openhands.integrations.service_types import UserGitInfo
-from openhands.server.settings import Settings
 from openhands.server.shared import server_config
-from openhands.storage.data_models.secrets import Secrets
 from openhands.utils.import_utils import get_impl
