+++ tests/unit/app_server/test_settings_api.py
@@ -6,8 +6,10 @@ from fastapi import Request
 from fastapi.testclient import TestClient
 from pydantic import SecretStr
 
+from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.app_server.secrets.secrets_store import SecretsStore
 from openhands.app_server.settings.file_settings_store import FileSettingsStore
+from openhands.app_server.settings.settings_models import Settings
 from openhands.app_server.settings.settings_store import SettingsStore
 from openhands.integrations.provider import ProviderToken, ProviderType
 from openhands.integrations.service_types import UserGitInfo
@@ -19,8 +21,6 @@ from openhands.sdk.settings import (
 )
 from openhands.server.app import app
 from openhands.server.user_auth.user_auth import UserAuth
-from openhands.storage.data_models.secrets import Secrets
-from openhands.storage.data_models.settings import Settings
 from openhands.storage.memory import InMemoryFileStore
 
 _EXPOSE = {'expose_secrets': True}
