+++ tests/unit/mcp/test_mcp_integration.py
@@ -5,11 +5,11 @@ from unittest.mock import AsyncMock, patch
 import pytest
 
 from openhands.app_server.settings.file_settings_store import FileSettingsStore
+from openhands.app_server.settings.settings_models import Settings
 from openhands.core.config.mcp_config import MCPConfig, RemoteMCPServer
 from openhands.sdk.llm import LLM
 from openhands.sdk.settings import AgentSettings
 from openhands.server.user_auth.default_user_auth import DefaultUserAuth
-from openhands.storage.data_models.settings import Settings
 
 
 def _sdk_mcp_config(settings: Settings) -> MCPConfig | None:
