+++ tests/unit/storage/data_models/test_secret_store.py
@@ -5,12 +5,12 @@ from typing import Any
 
 from pydantic import SecretStr
 
+from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.integrations.provider import (
     CustomSecret,
     ProviderToken,
     ProviderType,
 )
-from openhands.storage.data_models.secrets import Secrets
 
 
 class TestSecrets:
