+++ tests/unit/storage/data_models/test_settings.py
@@ -5,6 +5,7 @@ import pytest
 from fastmcp.mcp_config import MCPConfig
 from pydantic import SecretStr
 
+from openhands.app_server.settings.settings_models import Settings
 from openhands.core.config.llm_config import LLMConfig
 from openhands.core.config.openhands_config import OpenHandsConfig
 from openhands.core.config.sandbox_config import SandboxConfig
@@ -16,7 +17,6 @@ from openhands.sdk.settings import (
     ConversationSettings,
 )
 from openhands.sdk.settings.model import CondenserSettings, VerificationSettings
-from openhands.storage.data_models.settings import Settings
 
 
 def test_settings_from_config():
@@ -38,7 +38,7 @@ def test_settings_from_config():
     )
 
     with patch(
-        'openhands.storage.data_models.settings.load_openhands_config',
+        'openhands.app_server.settings.settings_models.load_openhands_config',
         return_value=mock_app_config,
     ):
         settings = Settings.from_config()
@@ -73,7 +73,7 @@ def test_settings_from_config_no_api_key():
     )
 
     with patch(
-        'openhands.storage.data_models.settings.load_openhands_config',
+        'openhands.app_server.settings.settings_models.load_openhands_config',
         return_value=mock_app_config,
     ):
         settings = Settings.from_config()
